requirements.txt
amqp==2.3.2
billiard==3.5.0.4
blinker==1.4
celery==4.2.1
Click==7.0
Flask==1.0.2
Flask-Mail==0.9.1
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
python-dotenv==0.9.1
pytz==2018.5
redis==2.10.6
vine==1.1.4
Werkzeug==0.14.1
